Rapid Response Systems - Standardized data elements (PDF)
Team Activations and Interventions
When studying the potential benefit of an MET or other rapid response system, the cost of providing a team response, measured in terms of time that team members are absent from their usual duties, may be an important consideration. Therefore, collection and reporting of the time of team activation and the time of completion of activation are core data elements. It is widely reported that time reporting that is not centralized or adjusted to a single clock is unreliable, and thus, hospitals that adjust to a single timepiece should be able to compare time-response effects. The interventions provided by the team are crucial for understanding the range and type of interventions that may be required to treat patients with deteriorating conditions. These interventions may also have an impact on patient outcome. The MET commonly undertakes therapies such as providing supplementary oxygen and fluids. These simple interventions may be as important in determining outcome as more invasive therapies.33

After team activation, team members may also consider institution of treatment-limitation orders when escalation of medical support is considered futile or inappropriate. These interventions, or the opening of discussion about treatment limitations, are an important part of response team activities and constitute an indicator of quality of care. In particular, the institution of a do-not-attempt-resuscitation (DNAR)/not-for-resuscitation (NFR) order is an important treatment. Institution of any treatment-limitation order or DNAR/NFR order should be collected and reported in order to understand the contribution of response teams to patient management.

Patient Outcome Measures
Patient outcomes are the most important measures of the effect of a system-wide intervention. Outcomes should be collected accurately to enable quality assurance of existing care and research into the different types of system responses to define future standards of care and to allow comparison between centers. The listed outcomes have been chosen for their simplicity and their importance in understanding system interventions. In particular, the patient’s vital status at time of discharge was selected to enable comparison of rates of serious adverse outcome, including cardiac arrest, because these are central indicators of patient outcome.

Hospital Outcome Measures
The hospital-level data elements related to DNAR/NFR deaths have been defined as core data in recognition of the potential for these system interventions to improve quality of care, particularly for patients at the end of life. It is also important to track the number of DNAR/NFR deaths when the cardiac arrest event rate and other indicators of system effectiveness are being monitored. This will help determine whether any decrease in cardiac arrests is the result of an increase in the rate of assignment of DNAR/NFR orders, and if so, whether this is an appropriate response and what proportion of the reduction in cardiac arrests can be attributed to this change. We anticipate that many institutions may have some difficulty in collecting this information, because it is often recorded poorly or not reported at all, but we urge investigators to make every effort to include information about DNAR/NFR deaths and the location of these deaths in their data collection and reporting. The MET has the best chance of altering outcome for patients who die unexpectedly in areas where the level of care is inappropriately low, and this is where the attention of most of these teams is directed. Unexpected deaths that occur in critical care areas, operating rooms, and emergency departments are not likely to be affected by a system such as the MET, because the level of care should already be optimal.